requirements.txt
mariadb==1.1.6
mysql_connector_repackaged==0.3.1
paramiko==3.0.0
psycopg2==2.9.5
python-dotenv==1.0.0
sshtunnel==0.4.0
